Fix for pip stripping "-e" from command when local source is not VCS. (#3426)
Code checked that `name` contained a VCS extension, AND that `editable: yes`. If one
was false, then the "-e" argument was not applied. I think (and believe this was the
original author's intent based on comments) that "-e" should be applied if either case
was true. This commit changes the boolean check from *AND* to *OR*.